+++ src/version-hooks/maven.ts
@@ -1,6 +1,6 @@
 import type {ExecOptions} from 'child_process';
 import {existsSync} from 'fs';
-import path from 'path';
+import {resolve, sep as pathSeparator} from 'path';
 
 import type {ReleaseContext, VersionFunction} from '../types';
 
@@ -9,7 +9,7 @@ import {Command} from './command';
 export function MavenPom(options?: ExecOptions): VersionFunction {
 
     const mavenPomCmd = (context: ReleaseContext) => {
-        const mvn = existsSync(path.resolve(context.directory, 'mvnw')) ? `.${path.sep}mvnw` : 'mvn';
+        const mvn = existsSync(resolve(context.directory, 'mvnw')) ? `.${pathSeparator}mvnw` : 'mvn';
         return `${mvn} versions:set -DnewVersion="${context.version.version}" -DgenerateBackupPoms=false`;
     };