+++ src/version-hooks/write-file.ts
@@ -1,6 +1,6 @@
 import {writeFile as writeFileCallback, WriteFileOptions} from 'fs';
 import {promisify} from 'util';
-import path from 'path';
+import {resolve} from 'path';
 
 import type {ReleaseContext, VersionFunction} from '../types';
 
@@ -13,7 +13,7 @@ export function WriteFile(
 ): VersionFunction {
     function pluginFunction(context: ReleaseContext): Promise<void> {
         const fileContent = content ? content(context) : context.version.version;
-        const fullFilePath = path.resolve(context.directory, file);
+        const fullFilePath = resolve(context.directory, file);
         return writeFile(fullFilePath, fileContent, writeFileOptions);
     }